requirements.txt
asgiref==3.11.0
black==25.12.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
Django==6.0.1
django-cors-headers==4.9.0
django-environ==0.12.0
django-filter==25.2
django-stubs==5.2.8
django-stubs-ext==5.2.8
djangorestframework==3.16.1
djangorestframework-camel-case==1.4.2
djangorestframework-stubs==3.16.6
drf-yasg==1.21.11
idna==3.11
inflection==0.5.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==1.0.2
platformdirs==4.5.1
psycopg==3.3.2
pytokens==0.3.0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
sqlparse==0.5.5
types-PyYAML==6.0.12.20250915
types-requests==2.32.4.20260107
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
